+++ services/workflows-service/src/providers/file/file-provider/local-file.service.ts
@@ -6,6 +6,7 @@ import * as path from 'path';
 
 export class LocalFileService implements IFileProvider {
   protected client;
+
   constructor(...args: any) {
     this.client = fs;
   }
@@ -42,14 +43,16 @@ export class LocalFileService implements IFileProvider {
       const distFilePath = this.__removeFilePrefix(remoteFileConfig as TRemoteUri);
 
       const readStream = this.client.createReadStream(this.__removeFilePrefix(localFilePath));
-      const writeStream = this.client.createWriteStream(this.__removeFilePrefix(distFilePath))
+      const writeStream = this.client.createWriteStream(this.__removeFilePrefix(distFilePath));
 
-      readStream.pipe(writeStream)
+      readStream
+        .pipe(writeStream)
         .on('finish', () => {
-        resolve(distFilePath);
-      }).on('error', err => {
-        reject(err);
-      });
+          resolve(distFilePath);
+        })
+        .on('error', err => {
+          reject(err);
+        });
     });
   }