+++ services/workflows-service/src/providers/file/file.service.ts
@@ -128,7 +128,6 @@ export class FileService {
 
     return {
       remoteFilePath,
-      localFilePath,
       mimeType: fileType?.mime,
     };
   }