+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ballerine/headless-example
 
+## 0.1.19
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.7.21
+  - @ballerine/workflow-browser-sdk@0.5.19
+
 ## 0.1.18
 
 ### Patch Changes